+++ app/components/messaging/channel-info.js
@@ -91,7 +91,7 @@ export default class ChannelInfo extends SafeComponent {
                             {tu('title_admin')}
                         </Text>
                     </View>}
-                    <Menu>
+                    {chat.canIAdmin && <Menu>
                         <MenuTrigger
                             renderTouchable={() => <TouchableOpacity pressRetentionOffset={vars.pressRetentionOffset} />}
                             style={{ padding: vars.iconPadding }}>
@@ -111,7 +111,7 @@ export default class ChannelInfo extends SafeComponent {
                                 <Text>{tx('button_remove')}</Text>
                             </MenuOption>
                         </MenuOptions>
-                    </Menu>
+                    </Menu>}
                 </View>
             </View>
         );
@@ -137,12 +137,13 @@ export default class ChannelInfo extends SafeComponent {
 
     renderThrow() {
         const chat = chatState.currentChat;
+        const invited = chatState.chatInviteStore.sent.get(chat.id);
         const body = (
             <View>
                 {this.lineBlock(this.topicTextBox())}
-                {this.lineBlock(this.action(tx('button_leaveChannel'), 'remove-circle-outline', this.leaveChannel), true)}
-                {this.lineBlock(this.action(tx('button_deleteChannel'), 'delete', this.deleteChannel))}
-                {chat.participants && this.lineBlock(
+                {chat.canILeave && this.lineBlock(this.action(tx('button_leaveChannel'), 'remove-circle-outline', this.leaveChannel), true)}
+                {chat.canIAdmin && this.lineBlock(this.action(tx('button_deleteChannel'), 'delete', this.deleteChannel))}
+                {chat.joinedParticipants && this.lineBlock(
                     <View style={{ paddingVertical: 8 }}>
                         <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'space-between', flexGrow: 1 }}>
                             <Text style={[textStyle, { marginBottom: 12 }]}>
@@ -150,7 +151,17 @@ export default class ChannelInfo extends SafeComponent {
                             </Text>
                             {icons.dark('add-circle-outline', () => chatState.routerModal.channelAddPeople())}
                         </View>
-                        {chat.participants.map(this.participant)}
+                        {chat.joinedParticipants.map(this.participant)}
+                    </View>
+                )}
+                {invited && this.lineBlock(
+                    <View style={{ paddingVertical: 8 }}>
+                        <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'space-between', flexGrow: 1 }}>
+                            <Text style={[textStyle, { marginBottom: 12 }]}>
+                                {tx('title_invited')}
+                            </Text>
+                        </View>
+                        {invited.map(this.participant)}
                     </View>
                 )}
             </View>